updating theme, imports and icon styling props
