--- .eslintrc
@@ -1,10 +1,0 @@
-{
-    "parser": "babel-eslint",
-    "plugins": [
-        "@emotion"
-    ],
-    "rules": {
-        "@emotion/pkg-renaming": "error",
-        "@emotion/styled-import": "error"
-    }
-}
