+++ package.json
@@ -42,7 +42,7 @@
         "babel-preset-react-app": "^9.1.2",
         "identity-obj-proxy": "^3.0.0",
         "jest": "^26.4.2",
-        "jest-emotion": "^10.0.32",
+        "@emotion/jest": "^11.1.0",
         "prettier": "^2.1.2",
         "react": "^16.13.1",
         "react-dom": "^16.13.1",
@@ -64,12 +64,8 @@
     ],
     "dependencies": {
         "@emotion/babel-preset-css-prop": "^11.0.0",
-        "@emotion/core": "^11.0.0",
-        "@emotion/eslint-plugin": "^11.0.0",
-        "@emotion/jest": "^11.1.0",
         "@emotion/react": "^11.1.4",
         "babel-eslint": "^10.1.0",
-        "emotion-theming": "^10.0.27",
         "eslint": "^7.17.0",
         "react-dnd": "^11.1.3",
         "react-dnd-html5-backend": "^11.1.3",
