+++ src/components/__private/componentFooter.styles.tsx
@@ -1,6 +1,5 @@
-import { css } from '@emotion/react';
+import { css, Theme } from '@emotion/react';
 import { Colors } from '../../shared/constants/colors';
-import { Theme } from '../theme/theme.types';
 
 export const disabledSpan = (theme: Theme) => css`
     box-sizing: border-box;
