+++ src/components/__private/componentFooter.tsx
@@ -1,12 +1,9 @@
 /** @jsx jsx */
-import { jsx, ThemeProvider } from '@emotion/react';
-import { useTheme } from '@emotion/react';
+import { jsx, useTheme } from '@emotion/react';
 import React from 'react';
 
-import { Theme } from '../theme/theme.types';
 import { ComponentFooterProps } from './componentFooter.types';
 import { disabledSpan, errorSpan } from './componentFooter.styles';
-import ThemeWrapper from '../themeWrapper/themeWrapper';
 
 /**
  * Renders a <ComponentFooter /> component
@@ -24,23 +21,21 @@ const ComponentFooter: React.FC<ComponentFooterProps> = ({
     //const localTheme: Theme;
     const theme = useTheme();
     return (
-        <ThemeWrapper>
-            <React.Fragment>
-                {disabled && disabledText && (
-                    <span css={disabledSpan(theme)}>{disabledText}</span>
-                )}
-                {error && typeof error === 'string' && (
-                    <span css={errorSpan(theme)}>{error}</span>
-                )}
-                {error && Array.isArray(error) && (
-                    <span css={errorSpan(theme)}>
-                        {error.reduce((a: string, b: string) => {
-                            return a.concat(', ').concat(b);
-                        })}
-                    </span>
-                )}
-            </React.Fragment>
-        </ThemeWrapper>
+        <React.Fragment>
+            {disabled && disabledText && (
+                <span css={disabledSpan(theme)}>{disabledText}</span>
+            )}
+            {error && typeof error === 'string' && (
+                <span css={errorSpan(theme)}>{error}</span>
+            )}
+            {error && Array.isArray(error) && (
+                <span css={errorSpan(theme)}>
+                    {error.reduce((a: string, b: string) => {
+                        return a.concat(', ').concat(b);
+                    })}
+                </span>
+            )}
+        </React.Fragment>
     );
 };
