+++ src/components/button/button.styles.tsx
@@ -1,7 +1,6 @@
-import { css } from '@emotion/react';
+import { css, Theme } from '@emotion/react';
 import { ButtonType } from './buttonType';
 import { Colors } from '../../shared/constants/colors';
-import { Theme } from '../theme/theme.types';
 
 export const button = (
     buttonType: ButtonType,
