+++ src/components/card/card.tsx
@@ -1,9 +1,7 @@
 /** @jsx jsx */
-import { jsx } from '@emotion/react';
+import { jsx, useTheme } from '@emotion/react';
 import React from 'react';
-import { useTheme } from '@emotion/react';
 
-import { Theme } from '../theme/theme.types';
 import cardProps from './card.types';
 import { card } from './card.styles';
 import { Colors } from '../../shared/constants/colors';
@@ -17,7 +15,7 @@ import { Colors } from '../../shared/constants/colors';
 
 const Card = React.forwardRef<HTMLInputElement, cardProps>(
     ({ children, backgroundColor = Colors.White, customCss }, ref) => {
-        const theme = useTheme<Theme>();
+        const theme = useTheme();
         const cardStyles = card(theme, backgroundColor);
         return (
             <div
