+++ src/components/draggable/DraggableButton styles.tsx
@@ -1,4 +1,4 @@
-import { css } from '@emotion/react';
+import { css, Theme } from '@emotion/react';
 export const draggableButton = () => css`
     display: flex;
     cursor: 'move';
