+++ src/components/dropdown/dropdown.styles.tsx
@@ -1,6 +1,5 @@
-import { css } from '@emotion/react';
+import { css, Theme } from '@emotion/react';
 import { Colors } from '../../shared/constants/colors';
-import { Theme } from '../theme/theme.types';
 import { ComponentWidth } from '../input';
 
 export const mainContainer = (theme: Theme, size: ComponentWidth) => css`
