+++ src/components/sidebar/Sidebar.tsx
@@ -1,13 +1,11 @@
 /** @jsx jsx */
-import { jsx } from '@emotion/react';
+import { jsx, useTheme } from '@emotion/react';
 import React from 'react';
-import { useTheme } from '@emotion/react';
 
-import { Theme } from '../theme/theme.types';
 import SidebarProps from './Sidebar.types';
 import { sidebar } from './Sidebar.styles';
 const Sidebar: React.FC<SidebarProps> = ({ children, visible = true }) => {
-    const theme = useTheme<Theme>();
+    const theme = useTheme();
     return (
         <React.Fragment>
             {visible && (
