--- src/components/theme/theme.types.ts
@@ -1,51 +1,0 @@
-export interface Theme {
-    fonts: {
-        robotoBold: string;
-        robotoRegular: string;
-        robotoMedium: string;
-    };
-    fontSizes: {
-        buttonText: string;
-        buttonTextWithSubtext: string;
-        buttonSubtext: string;
-        h1: string;
-        h1Mobile: string;
-        h2: string;
-        h2Mobile: string;
-        h3: string;
-        h4: string;
-        pSmall: string;
-        pSmallMobile: string;
-        pMedium: string;
-        pMediumMobile: string;
-        pLarge: string;
-        pLargeMobile: string;
-        input: string;
-        inputLabel: string;
-        inputDetail: string;
-        inputDisabled: string;
-        inputError: string;
-        tooltip: string;
-        textarea: string;
-        textareaLabel: string;
-        textareaDetail: string;
-        textareaDisabled: string;
-        textareaError: string;
-        dropdownList: string;
-        toggleDetail: string;
-        checkboxDetail: string;
-    };
-    breakpoints: {
-        sm: number;
-        md: number;
-    };
-    columns: {
-        column: number;
-        gutter: number;
-        twoColumns: number;
-        threeColumns: number;
-        fourColumns: number;
-        fiveColumns: number;
-        sixColumns: number;
-    };
-}
