+++ src/components/themeWrapper/themeWrapper.tsx
@@ -1,9 +1,7 @@
 /** @jsx jsx */
-import { jsx } from '@emotion/react';
+import { jsx, CacheProvider, ThemeProvider } from '@emotion/react';
 import React from 'react';
-import { CacheProvider } from '@emotion/react';
 import createCache from '@emotion/cache';
-import { ThemeProvider } from '@emotion/react';
 import createExtraScopePlugin from 'stylis-plugin-extra-scope';
 
 import defaultTheme from '../../shared/themes/theme';
@@ -12,6 +10,7 @@ import GlobalStyles from '../Layout/GlobalStyles/GlobalStyles';
 
 const STRONG_ID = 'trova-components-scope';
 const cache = createCache({
+    key: 'my-prefix-key',
     stylisPlugins: [createExtraScopePlugin(`#${STRONG_ID}`)],
 });
 
@@ -22,9 +21,7 @@ const cache = createCache({
  * @param  props.children - The elements to render inside the theme provider
  */
 
-const ThemeWrapper: React.FC<themeWrapperProps> = ({
-    children,
-}) => (
+const ThemeWrapper: React.FC<themeWrapperProps> = ({ children }) => (
     <CacheProvider value={cache}>
         <div id={STRONG_ID}>
             <GlobalStyles />
