+++ src/components/typography/h2.tsx
@@ -1,14 +1,12 @@
 /** @jsx jsx */
-import { jsx } from '@emotion/react';
+import { jsx, useTheme } from '@emotion/react';
 import React from 'react';
-import { useTheme } from '@emotion/react';
 
-import { Theme } from '../theme/theme.types';
 import { h2Style } from './h2.styles';
 import Typography from './typography.types';
 
 const H2: React.FC<Typography> = ({ children, ...otherProps }) => {
-    const theme = useTheme<Theme>();
+    const theme = useTheme();
     return (
         <h2 data-testid="h2" css={h2Style(theme)} {...otherProps}>
             {children}
