+++ src/components/typography/p.tsx
@@ -1,15 +1,17 @@
 /** @jsx jsx */
-import { jsx } from '@emotion/react';
+import { jsx, useTheme } from '@emotion/react';
 import React from 'react';
-import { useTheme } from '@emotion/react';
 
-import { Theme } from '../theme/theme.types';
 import { pStyle } from './p.styles';
 import Typography from './typography.types';
 import { ParagraphSize } from './paragraphSize';
 
-const P: React.FC<Typography> = ({size = ParagraphSize.Medium, children, ...otherProps }) => {
-    const theme = useTheme<Theme>();
+const P: React.FC<Typography> = ({
+    size = ParagraphSize.Medium,
+    children,
+    ...otherProps
+}) => {
+    const theme = useTheme();
     return (
         <p data-testid="p" css={pStyle(size, theme)} {...otherProps}>
             {children}
